\name{dna.obj}
\alias{dna.obj}
\docType{data}
\title{Example DNA sequence data}
\description{
An example object of the class \code{\link{Dna}}. 
}
\usage{data(dna.obj)}
\format{
    \code{dna.obj} contains a \code{\link{Dna}} object. 
    }

\examples{
data(dna.obj)
dna.obj
}

\keyword{DNA ANALYSIS}
\keyword{DATASETS}
