\name{haplotypes-internal}
\alias{add.miss}
\alias{alltest}
\alias{as.sequence}
\alias{calcclust}
\alias{clsfyindels}
\alias{clustersG}
\alias{contr}
\alias{distance2}
\alias{ext.nodes}
\alias{fillendgaps}
\alias{fillna}
\alias{geodist.corr}
\alias{haplo}
\alias{has.cycle}
\alias{indtest}
\alias{multipolypoz}
\alias{pairneidist}
\alias{pairPhiSTdist}
\alias{parsnet}
\alias{pegas.amova}
\alias{polys}
\alias{recdistG}
\alias{ren.dupli}
\alias{ri}
\alias{sic}
\alias{steplimit}
\alias{tryconnect}
\alias{tryconnectG}
\alias{validity}
\alias{xhaps}
\alias{sic}
\alias{validity}
\alias{initialize-methods}
\alias{initialize,Dna-method}
\alias{initialize,Haplotype-method}
\alias{initialize,Parsimnet-method}
\title{Internal function(s)}
\description{Internal function(s)}
\keyword{internal}