\name{haplotypes-internal}
\alias{alltest}
\alias{calcclust}
\alias{clustersG}
\alias{fillendgaps}
\alias{fillna}
\alias{haplo}
\alias{pairneidist}
\alias{parsnet}
\alias{recdistG}
\alias{sic}
\alias{steplimit}
\alias{tryconnect}
\alias{tryconnectG}
\alias{validity}
\alias{xhaps}
\alias{sic}
\alias{initialize-methods}
\alias{initialize,Dna-method}
\alias{initialize,Haplotype-method}
\alias{initialize,Parsimnet-method}
\title{Internal function(s)}
\description{Internal function(s)}
\keyword{internal}